Make from_json a classmethod on DataFrame.
Deinining from_json at the module level forces the dataframe returned to be of
class rc.DataFrame, which isn't ideal if you're subclassing the dataframe to add
functionality to it.  Making it a classmethod allows dataframes loaded with
from_json to be of whatever class the method is called from.